던전 제작으로 인한 exp 받기
레벨업 브랜치 사용했어야했는데 안바꾸고 여기다가 해버렸네요. 죄송합니다.

+++ Text_RPG_Team/Monster.cs
@@ -100,9 +100,9 @@ namespace Text_RPG_Team
         //변수 조작 함수
 
         //몬스터 피격 시
-        public void TakeDamage(int damage)
+        public void TakeDamage(float damage)
         {
-            health -= damage;
+            health -= (int)damage;
             if (health <= 0)
             {
                 health = 0;
